[OpenReg] Add Develop Notes for Integrating New Backend into PyTorch
To facilitate the integration of the new backend, we plan to publish a new development note that details all the key components,
hoping to speed up the development of other accelerators.

This PR is the beginning of this note, and involve the part of registration of operators and we will gradually improve it and
keep in sync with OpenReg's code.

ghstack-source-id: d1239f31196dce2abb550590f193a23c5368dd50
Pull-Request-resolved: https://github.com/pytorch/pytorch/pull/158644